requirements.txt
# Core dependencies
playwright==1.40.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
mysql-connector-python==8.2.0
pymysql==1.1.0

# Task scheduling
celery==5.3.4
redis==5.0.1

# Dashboard
streamlit==1.29.0
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
